more refactorings and properly loading map for jojo

+++ src/ecs/component/agentinputcomponent.h
@@ -2,8 +2,8 @@
 
 struct AgentInputComponent
 {
-    int x_axis = 0;
-    int y_axis = 0;
+    float x_axis = 0;
+    float y_axis = 0;
     bool attack_requested = false;
     bool jump_requested = false;
     bool jump_just_requested = false;
